[refactor] Append 'OnDb' to repository implementation classes to clarify usage of database technology in class names
