+++ app/infrastructure/db/repositories/ExamResultRepositoryImplOnDb.scala
@@ -1,35 +1,38 @@
 package infrastructure.db.repositories
 
 import javax.inject.{Inject, Singleton}
-import domain.exam.entity.Exam
-import dto.infrastructure.exam.entity.ExamDto
 import infrastructure.db.DatabaseConfig
-import infrastructure.db.table.ExamTable
-import usecases.exam.repository.ExamRepository
+import infrastructure.db.table.ExamResultTable
+import scala.concurrent.{ExecutionContext, Future}
+import usecases.examResult.repository.ExamResultRepository
+import domain.examResult.entity.ExamResult
+import domain.exam.valueObject._
+import dto.infrastructure.examResult.entity.ExamResultDto
 import utils.Retry
-import scala.concurrent.{Future, ExecutionContext}
 import scala.concurrent.duration._
 import org.apache.pekko.actor.Scheduler
 import java.sql.SQLTransientConnectionException
 import java.time.ZonedDateTime
-import domain.exam.valueObject.ExamId
+import domain.examResult.valueObject.ExamResultId
 
 @Singleton
-class ExamRepositoryImpl @Inject() (
+class ExamResultRepositoryImplOnDb @Inject() (
     dbConfig: DatabaseConfig,
     scheduler: Scheduler
 )(implicit
     ec: ExecutionContext
-) extends ExamRepository {
+) extends ExamResultRepository {
   import dbConfig._
   import profile.api._
 
-  override def save(exam: Exam): Future[Either[String, Exam]] = {
-    val dto = ExamDto.fromDomain(exam)
-    val query = ExamTable.exams += dto
+  override def save(
+      examResult: ExamResult
+  ): Future[Either[String, ExamResult]] = {
+    val dto = ExamResultDto.fromDomain(examResult)
+    val query = ExamResultTable.examResults += dto
     Retry
       .withRetry(run(query), 3, 1.second)(ec, scheduler)
-      .map(_ => Right(exam))
+      .map(_ => Right(examResult))
       .recover {
         case ex: SQLTransientConnectionException =>
           Left("Database connection error")
@@ -38,19 +41,19 @@ class ExamRepositoryImpl @Inject() (
   }
 
   override def findById(
-      examId: ExamId
-  ): Future[Either[String, Option[Exam]]] = {
-    val query = ExamTable.exams
-      .filter(_.examId === examId.value)
+      examResultId: ExamResultId
+  ): Future[Either[String, Option[ExamResult]]] = {
+    val query = ExamResultTable.examResults
+      .filter(_.examResultId === examResultId.value)
       .result
       .headOption
     Retry
       .withRetry(run(query), 3, 1.second)(ec, scheduler)
       .map {
         case Some(dto) =>
-          ExamDto.toDomain(dto) match {
-            case Right(exam) => Right(Some(exam))
-            case Left(error) => Left(error)
+          ExamResultDto.toDomain(dto) match {
+            case Right(examResult) => Right(Some(examResult))
+            case Left(error)       => Left(error)
           }
         case None => Right(None)
       }
@@ -61,26 +64,21 @@ class ExamRepositoryImpl @Inject() (
       }
   }
 
-  override def findByDueDate(
-      startDate: ZonedDateTime,
-      endDate: ZonedDateTime
-  ): Future[Either[String, Seq[Exam]]] = {
-    val query = ExamTable.exams
-      .filter(exam => exam.dueDate >= startDate && exam.dueDate <= endDate)
+  override def findByExamId(
+      examId: ExamId
+  ): Future[Either[String, Seq[ExamResult]]] = {
+    val query = ExamResultTable.examResults
+      .filter(_.examId === examId.value)
       .result
-
     Retry
       .withRetry(run(query), 3, 1.second)(ec, scheduler)
       .map { dtos =>
-        val exams = dtos.map(ExamDto.toDomain)
-        val (errors, validExams) = exams.partitionMap(identity)
-
+        val examResults = dtos.map(ExamResultDto.toDomain)
+        val errors = examResults.collect { case Left(error) => error }
         if (errors.nonEmpty) {
           Left(errors.mkString(", "))
-        } else if (validExams.isEmpty) {
-          Left("No exams found for the given period")
         } else {
-          Right(validExams)
+          Right(examResults.collect { case Right(examResult) => examResult })
         }
       }
       .recover {
@@ -90,13 +88,16 @@ class ExamRepositoryImpl @Inject() (
       }
   }
 
-  override def update(exam: Exam): Future[Either[String, Exam]] = {
-    val dto = ExamDto.fromDomain(exam)
-    val query =
-      ExamTable.exams.filter(_.examId === dto.examIdDto.value).update(dto)
+  override def update(
+      examResult: ExamResult
+  ): Future[Either[String, ExamResult]] = {
+    val dto = ExamResultDto.fromDomain(examResult)
+    val query = ExamResultTable.examResults
+      .filter(_.examResultId === dto.examResultId)
+      .update(dto)
     Retry
       .withRetry(run(query), 3, 1.second)(ec, scheduler)
-      .map(_ => Right(exam))
+      .map(_ => Right(examResult))
       .recover {
         case ex: SQLTransientConnectionException =>
           Left("Database connection error")
